scope user card search in sales to the context itself
